💈 Clean things up, Get things working
