+++ api/plugins/multiSensor/Readme.md
@@ -1,7 +1,7 @@
 # Multi Sensor Plugin
 
-To use this plugin's email alerts you will need to install the Nodemailer package
+To use this plugin you will need to install some extra packages
 Run this command in the api directory:
 ```bash
-$ npm i nodemailer
+$ npm i nodemailer axios
 ```
