+++ api/plugins/multiSensor/common.js
@@ -1,4 +1,7 @@
+const common = require('../../src/common');
+
 // Functions / Constants used my multiple files for this plugin
+// Or just things I want to move out on index.js
 
 /**
  * Types of Alerts
@@ -17,6 +20,41 @@ const Type = {
     outOf: 'outOf'
 };
 
+/**
+ * Capatilize the first letter of each word
+ * @param {string} name The name to capitalize
+ * @returns {string} The capitalized name
+ */
+function niceName(name) {
+    let working = name.toLowerCase().split(' ');
+    working.forEach((e, i) => {
+        working[i] = e[0].toUpperCase() + e.substr(1);
+    });
+    return working.join(' ');
+}
+
+/**
+ * Get the data from backend server
+ * @param {Object} config Global Config
+ * @returns {Promise} The temp data from the backend server
+ * @async
+ */
+function getData(config) {
+    let sen = `http://${config.sensor.ip}:${config.sensor.port}`;
+    return new Promise((resolve, reject) => {
+        common
+            .get(`${sen}/temp`)
+            .then(data => {
+                data = JSON.parse(data);
+                lastData = data;
+                resolve(data);
+            })
+            .catch(reject);
+    });
+}
+
 module.exports = {
-    Type
+    Type,
+    niceName,
+    getData
 };
